Move the contacts under Directory and drop the standalone screen
