--- source/features/contacts/types.ts
@@ -1,10 +1,0 @@
-export type ContactType = {
-	title: string
-	phoneNumber?: string
-	buttonText: string
-	buttonLink?: string
-	category: string
-	image?: string
-	text: string
-	synopsis: string
-}
