--- uitests/ModuleImportantContactsTests.swift
@@ -1,9 +1,0 @@
-import XCTest
-
-class ModuleImportantContactsTests: UITestCase {
-	func testIsReachableFromHomescreen() throws {
-		ImportantContactsScreen(app: app)
-			.navigate()
-			.verifyImportantContactsTitle()
-	}
-}
